requirements.txt
Flask==2.3.3
Pillow==10.0.0
numpy==1.24.3
gunicorn==21.2.0